changed double value annualInterestRate

+++ src/ClassesInterfaces/CalculateInterest.java
@@ -1,5 +1,5 @@
 package ClassesInterfaces;
 
 public interface CalculateInterest {
-    int calculateMonthlyInterest();
+    double calculateMonthlyInterest();
 }
